requirements.txt
pymongo
python-telegram-bot
spacy
idiomatch
lemminflect
python-dotenv
pyenchant